fix(demo-prover-avail): minor fixed
