+++ examples/avail-helper/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "demo-sequencer-avail"
+name = "avail-helper"
 version = "0.1.0"
 edition = "2021"
