+++ examples/demo-prover-avail/host/Dockerfile
