+++ examples/demo-prover-avail/host/rollup_config.toml
@@ -1,10 +1,10 @@
 # We define the rollup's genesis to occur at Avail block number `start_height`. The rollup will ignore
 # any Avail blocks before this height
-start_height = 7687
+start_height = 147699
 
 [da]
 light_client_url = "http://127.0.0.1:7000"
-node_client_url = "ws://127.0.0.1:9944"
+node_client_url = "wss://kate.avail.tools:443/ws"
 
 [runner.storage]
 # The path to the rollup's data directory. Paths that do not begin with `/` are interpreted as relative paths.
