+++ examples/demo-prover-avail/host/src/main.rs
@@ -106,7 +106,7 @@ async fn main() -> Result<(), anyhow::Error> {
         );
 
         let filtered_block = da_service.get_finalized_at(height).await?;
-        let header_hash = hex::encode(filtered_block.header.header.hash());
+        let header_hash = hex::encode(filtered_block.hash());
         host.write_to_guest(&filtered_block.header);
         let (blob_txs, inclusion_proof, completeness_proof) =
             da_service.extract_relevant_txs_with_proof(&filtered_block);
